Update to upstream version '259~rc2'
with Debian dir cf601b8ede8f71b078792ce228102fe98830a5fc

+++ src/shared/reread-partition-table.c
@@ -14,7 +14,6 @@
 #include "errno-util.h"
 #include "fd-util.h"
 #include "log.h"
-#include "parse-util.h"
 #include "reread-partition-table.h"
 #include "set.h"
 #include "string-util.h"
@@ -180,7 +179,7 @@ static int process_partition(
                  * just to make a point. */
                 partition = sd_device_unref(partition);
 
-                r = block_device_remove_partition(fd, subnode, (int) nr);
+                r = block_device_remove_partition(fd, subnode, nr);
                 if (r < 0)
                         return log_device_debug_errno(d, r, "Failed to remove kernel partition device '%s' in order to recreate it: %m", subnode);
